feat：重构文件管理界面

+++ huasen-frontend/admin/src/views/file-manage/FileManage.vue
@@ -8,46 +8,64 @@
 
 <template>
   <div class="FileManage">
-    <FileUpLoad @upload="upload"></FileUpLoad>
-    <main>
-      <ul v-if="this.files.length !== 0">
-        <li class="file-card" v-for="(item, index) in displayFiles" :key="`${item}-${index}`">
-          <div class="content-panel">
-            <div class="file-box">
-              <i v-if="!judgeImgFile(item)" class="iconfont" :class="getFileIcon(item)"></i>
-              <img v-else :src="item" />
+    <FileUpLoad class="relative z-10" @upload="upload"></FileUpLoad>
+    <main class="px-px-24 relative z-0">
+      <div class="head bg-gray-0 flex items-center -mt-px-10 py-px-14">
+        <div class="text-lg flex items-center text-gray-800">
+          <!-- <i class="el-icon-receiving text-gray-700 text-22px mr-px-4"></i> -->
+          文件列表
+        </div>
+        <el-popconfirm @confirm="remove('multiple')" class="ml-auto" popper-class="delete-popcomfirm" title="确定删除吗？">
+          <el-button slot="reference" size="small" type="danger">批量删除</el-button>
+        </el-popconfirm>
+        <!-- <el-button size="small" type="info" @click="downloadStore">下载全部</el-button> -->
+      </div>
+      <el-table ref="fileTable" :data="displayFiles" border style="width: 100%">
+        <el-table-column type="selection" width="45"> </el-table-column>
+        <el-table-column type="index" label="序号" width="65"> </el-table-column>
+        <el-table-column prop="url" label="文件地址"> </el-table-column>
+        <el-table-column prop="type" label="文件类型" width="100">
+          <template slot-scope="scope">
+            <font>{{ fileTypeName[scope.row.type] || '未知' }}</font>
+          </template>
+        </el-table-column>
+        <el-table-column prop="" label="缩略图" width="150">
+          <template slot-scope="scope">
+            <div class="w-full h-px-50 p-px-2 flex justify-center items-center rounded">
+              <i v-if="!scope.row.isImg" class="iconfont text-54px" :class="scope.row.iconClass"></i>
+              <img v-else class="max-w-full max-h-full" v-lazy :src="scope.row.url" />
             </div>
-          </div>
-          <div class="shadow-panel">
-            <span @click="copy(item, index)">复制</span>
-            <span @click="preview(item, index)">预览</span>
-            <el-popconfirm @confirm="remove(item, index)" popper-class="delete-popcomfirm" title="您确定删除该项吗？">
-              <span slot="reference">删除</span>
+          </template>
+        </el-table-column>
+        <el-table-column prop="date" label="操作" width="320">
+          <template slot-scope="scope">
+            <el-popconfirm @confirm="remove('single', scope.row.url, scope.$index)" class="mr-px-10" popper-class="delete-popcomfirm" title="确定删除吗？">
+              <el-button slot="reference" size="mini" type="danger">删 除</el-button>
             </el-popconfirm>
-          </div>
-        </li>
-        <el-pagination
-          class="w-full flex mt-px-28"
-          :page-sizes="[10, 20, 50, 100]"
-          :current-page="currentPage"
-          :pageSize="pageSize"
-          :total="total"
-          @size-change="handlePageSizeChange"
-          @current-change="handleCurrentPageChange"
-          popper-class="page-size-popper"
-          layout="total, sizes, prev, pager, next, jumper"
-          background
-        >
-        </el-pagination>
-      </ul>
-      <Empty v-else class="mt-px-64"></Empty>
+            <el-button size="mini" @click="copy(scope.row.url, scope.$index)" type="warning">复 制</el-button>
+            <el-button size="mini" @click="preview(scope.row.url, scope.$index)" type="primary">预 览</el-button>
+          </template>
+        </el-table-column>
+      </el-table>
+      <el-pagination
+        class="w-full flex mt-px-28"
+        :page-sizes="[10, 20, 50, 100]"
+        :current-page="currentPage"
+        :pageSize="pageSize"
+        :total="total"
+        @size-change="handlePageSizeChange"
+        @current-change="handleCurrentPageChange"
+        popper-class="page-size-popper"
+        layout="total, sizes, prev, pager, next, jumper"
+        background
+      >
+      </el-pagination>
     </main>
   </div>
 </template>
 
 <script>
 import FileUpLoad from '@/components/common/file-upload/FileUpload.vue';
-import Empty from '@/components/content/empty/Empty.vue';
 export default {
   name: 'FileManage',
   computed: {
@@ -64,23 +82,39 @@ export default {
     return {
       files: [],
 
+      fileTypeName: {
+        icon: '图标',
+        article: '文章附件',
+        default: '其它',
+        pdf: 'PDF',
+        zip: '压缩包',
+        banner: '封面',
+        img: '图库',
+        'open-sh': '开放脚本',
+      },
       type2icon: {
-        png: 'icon-tupian',
-        jpg: 'icon-tupian',
-        jpeg: 'icon-tupian',
-        zip: 'icon-yasuobao',
-        rar: 'icon-yasuobao',
-        pdf: 'icon-pdf',
-        md: 'icon-file-markdown',
-        doc: 'icon-file-word',
-        docx: 'icon-file-word',
-        xls: 'icon-excel',
-        xlsx: 'icon-excel',
-        ppt: 'icon-file-ppt',
-        pptx: 'icon-file-ppt',
+        png: 'icon-PNG',
+        jpg: 'icon-JPG',
+        jpeg: 'icon-JPEG',
+        zip: 'icon-ZIP',
+        rar: 'icon-RAR',
+        pdf: 'icon-PDF',
+        md: 'icon-MD',
+        doc: 'icon-DOC',
+        docx: 'icon-DOCX',
+        xls: 'icon-XLS',
+        xlsx: 'icon-XLSX',
+        ppt: 'icon-PPT',
+        pptx: 'icon-PPT',
         html: 'icon-HTML',
         css: 'icon-CSS',
-        js: 'icon-js',
+        js: 'icon-JS',
+        sh: 'icon-SH',
+        mp3: 'icon-MP3',
+        mp4: 'icon-MP4',
+        java: 'icon-JAVA',
+        jar: 'icon-JAR',
+        '7z': 'icon-SEVENZ',
       },
 
       // 前端分页相关
@@ -91,7 +125,6 @@ export default {
 
   components: {
     FileUpLoad,
-    Empty,
   },
 
   mounted() {
@@ -99,6 +132,10 @@ export default {
   },
 
   methods: {
+    downloadStore() {
+      this.API.downloadStoreByZip();
+    },
+
     async upload(file, index, callback) {
       let formdata = new FormData();
       formdata.append('file', file);
@@ -121,6 +158,7 @@ export default {
       return ['png', 'jpg', 'jpeg', 'svg', 'gif'].includes(ext);
     },
 
+    // 获取文件图标
     getFileIcon(filePath) {
       let ext = filePath
         .split('.')
@@ -129,13 +167,29 @@ export default {
       return this.type2icon[ext];
     },
 
-    copy(item, index) {
-      this.TOOL.copyTextToClip(item, '拷贝链接成功');
+    // 拷贝链接
+    copy(url, index) {
+      this.TOOL.copyTextToClip(url, '拷贝链接成功');
     },
 
-    remove(path, index) {
+    // 移除文件
+    remove(flag = 'multiple', url, index) {
+      let needRemoveUrls = [];
+      switch (flag) {
+        case 'multiple':
+          needRemoveUrls = this.$refs.fileTable.selection.map(item => {
+            return item.url;
+          });
+          break;
+        case 'single':
+          needRemoveUrls.push(url);
+          break;
+      }
+      debugger;
+      if (needRemoveUrls.length === 0) return;
+
       this.API.removeFile(
-        { filePath: path },
+        { filePaths: needRemoveUrls, isMultiple: true },
         {
           notify: true,
         },
@@ -144,6 +198,7 @@ export default {
       });
     },
 
+    // 预览并下载
     preview(item, index) {
       window.open(item);
     },
@@ -162,7 +217,20 @@ export default {
           notify: false,
         },
       ).then(res => {
-        this.files = res.data;
+        this.files = res.data.map(url => {
+          let params = url.split('/');
+          let type = params[1];
+          let ext = params[params.length - 1].split('.')[1];
+          let iconClass = this.getFileIcon(url);
+          let isImg = ['png', 'PNG', 'jpg', 'JPG', 'jpeg', 'JPEG', 'svg', 'SVG', 'gif', 'GIF'].includes(ext);
+          return {
+            url,
+            type,
+            ext,
+            iconClass,
+            isImg,
+          };
+        });
       });
     },
   },
